+++ code/app/src/main/java/com/castlemock/app/config/PropertyConfig.java
@@ -1,0 +1,13 @@
+package com.castlemock.app.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.PropertySource;
+
+@Configuration
+@PropertySource({
+        "classpath:file-repository.properties",
+        "classpath:spring.properties",
+        "classpath:swagger.properties",
+})
+public class PropertyConfig {
+}
