+++ code/app/src/main/resources/spring.properties
@@ -1,0 +1,4 @@
+spring.view.prefix=/WEB-INF/views/
+spring.view.suffix=.jsp
+spring.main.banner-mode=off
+spring.mvc.dispatch-options-request=true
