+++ code/app/src/main/resources/swagger.properties
@@ -1,0 +1,1 @@
+springfox.documentation.swagger.v2.path=/api/doc/rest/swagger
