+++ code/deploy/tomcat/jar/src/main/resources/application.properties
@@ -1,0 +1,9 @@
+app.name=CastleMock
+app.version=1.34-SNAPSHOT
+
+spring.main.banner-mode=off
+
+server.port=8080
+server.servlet.context-path=/castlemock
+server.mode.demo=false
+server.endpoint.address=
