+++ code/deploy/tomcat/war/src/main/resources/application.properties
@@ -1,0 +1,8 @@
+app.name=CastleMock
+app.version=1.34-SNAPSHOT
+
+spring.main.banner-mode=off
+
+server.port=8080
+server.mode.demo=false
+server.endpoint.address=
