--- code/web/app/src/main/java/com/castlemock/web/app/config/RestSecurityConfig.java
@@ -1,64 +1,0 @@
-/*
- * Copyright 2018 Karl Dahlgren
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.castlemock.web.app.config;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.core.annotation.Order;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
-import org.springframework.security.core.userdetails.UserDetailsService;
-
-/**
- * The class {@link RestSecurityConfig} provides the configuration for the REST interfaces.
- *
- * @author Karl Dahlgren
- * @since 1.25
- */
-@Configuration
-@Order(2)
-@EnableWebSecurity
-public class RestSecurityConfig extends WebSecurityConfigurerAdapter {
-
-    @Autowired
-    @Qualifier("userDetailsService")
-    private UserDetailsService userDetailsService;
-
-    /**
-     * The method configure is responsible for the security configuration.
-     *
-     * @param httpSecurity httpSecurity will be used to configure the authentication process.
-     * @throws Exception Throws an exception if the configuration fails
-     */
-    @Override
-    protected void configure(final HttpSecurity httpSecurity) throws Exception {
-        httpSecurity
-                .antMatcher("/api/rest/**")
-                    .authorizeRequests()
-                    .anyRequest()
-                    .authenticated()
-                .and()
-                    .httpBasic()
-                .and()
-                    .csrf().disable();
-
-        httpSecurity.headers().cacheControl().disable();
-    }
-
-}
