--- code/web/jar/src/main/resources/application.properties
@@ -1,83 +1,0 @@
-
-app.name=CastleMock
-app.version=1.34-SNAPSHOT
-
-spring.view.prefix=/WEB-INF/views/
-spring.view.suffix=.jsp
-spring.main.banner-mode=off
-spring.mvc.dispatch-options-request=true
-
-server.port=8080
-server.mode.demo=false
-server.endpoint.address=
-
-springfox.documentation.swagger.v2.path=/api/doc/rest/swagger
-
-token.validity.seconds=31536000
-token.file.directory=${base.file.directory}/token
-token.file.name=tokens.token
-
-base.file.directory=${user.home}/.castlemock
-configuration.file.directory=${base.file.directory}/configuration
-configuration.file.extension=.conf
-user.file.directory=${base.file.directory}/user
-user.file.extension=.user
-temp.file.directory=${base.file.directory}/tmpFiles
-
-soap.project.file.directory=${base.file.directory}/soap/project/v2
-soap.project.file.extension=.prj
-soap.port.file.directory=${base.file.directory}/soap/port/v2
-soap.port.file.extension=.port
-soap.operation.file.directory=${base.file.directory}/soap/operation/v2
-soap.operation.file.extension=.opr
-soap.response.file.directory=${base.file.directory}/soap/response/v2
-soap.response.file.extension=.res
-soap.resource.file.directory=${base.file.directory}/soap/resource/v2
-soap.resource.file.extension=.rsc
-
-soap.event.file.directory=${base.file.directory}/soap/event/v1
-soap.event.file.extension=.event
-soap.event.max=100
-
-rest.project.file.directory=${base.file.directory}/rest/project/v2
-rest.project.file.extension=.prj
-rest.application.file.directory=${base.file.directory}/rest/application/v2
-rest.application.file.extension=.app
-rest.resource.file.directory=${base.file.directory}/rest/resource/v2
-rest.resource.file.extension=.res
-rest.method.file.directory=${base.file.directory}/rest/method/v2
-rest.method.file.extension=.met
-rest.response.file.directory=${base.file.directory}/rest/response/v2
-rest.response.file.extension=.rest
-
-rest.event.file.directory=${base.file.directory}/rest/event/v1
-rest.event.file.extension=.event
-rest.event.max=100
-
-graphql.project.file.directory=${base.file.directory}/graphql/project/v1
-graphql.project.file.extension=.prj
-graphql.application.file.directory=${base.file.directory}/graphql/application/v1
-graphql.application.file.extension=.prj
-graphql.enumtype.file.directory=${base.file.directory}/graphql/enum/v1
-graphql.enumtype.file.extension=.enum
-graphql.mutation.file.directory=${base.file.directory}/graphql/mutation/v1
-graphql.mutation.file.extension=.mut
-graphql.object.file.directory=${base.file.directory}/graphql/object/v1
-graphql.object.file.extension=.obj
-graphql.query.file.directory=${base.file.directory}/graphql/query/v1
-graphql.query.file.extension=.que
-graphql.subscription.file.directory=${base.file.directory}/graphql/subscription/v1
-graphql.subscription.file.extension=.sub
-graphql.attribute.file.directory=${base.file.directory}/graphql/attribute/v1
-graphql.attribute.file.extension=.att
-graphql.event.file.directory=${base.file.directory}/graphql/event/v1
-graphql.event.file.extension=.event
-graphql.event.max=100
-
-
-
-legacy.soap.project.v1.directory=${base.file.directory}/soap/project
-legacy.rest.project.v1.directory=${base.file.directory}/rest/project
-legacy.soap.event.v1.directory=${base.file.directory}/soap/event
-legacy.rest.event.v1.directory=${base.file.directory}/rest/event
-legacy.soap.resource.v1.directory=${base.file.directory}/soap/resource
